fix: make it lazy
So, the default behaviour for workers is to load application into master
process and make as many forks as requested. The threads are created in
the master process and aren't forked into children. Not sure why memory
isn't shared as it should be copy-on-write. Anyway seems like it quite
safe to use --lazy.